# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Demo Days Off Calendar API Documentation",
        "description": "API Documentation for Demo Days Off Calendar using Swagger",
        "contact": {
            "email": "gradiph@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/calendar": {
            "get": {
                "tags": [
                    "Calendar"
                ],
                "summary": "Get Calendar for Authenticated User",
                "description": "Get calendar data for authenticated user",
                "operationId": "calendar.self",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": [
                                        {
                                            "date": "2025-06-17",
                                            "reason": "Odio quibusdam suscipit nemo repellendus fugit.",
                                            "approvedAt": null
                                        },
                                        {
                                            "date": "2025-05-23",
                                            "reason": "Beatae omnis optio dolor minus occaecati dolor.",
                                            "approvedAt": "2025-05-31 00:11:09"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Calendar"
                ],
                "summary": "Create New Days Off",
                "description": "Create New Days Off",
                "operationId": "calendar.create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "newOffDays",
                                    "userId",
                                    "reason"
                                ],
                                "properties": {
                                    "newOffDays": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "default": "2025-06-17"
                                        }
                                    },
                                    "userId": {
                                        "type": "number",
                                        "default": "1"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "default": "Lorem ipsum"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": []
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/calendar/{user_id}": {
            "get": {
                "tags": [
                    "Calendar"
                ],
                "summary": "Get Calendar for Specific User",
                "description": "Get calendar data for requested user",
                "operationId": "calendar.user",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": [
                                        {
                                            "date": "2025-06-17",
                                            "reason": "Odio quibusdam suscipit nemo repellendus fugit.",
                                            "approvedAt": null
                                        },
                                        {
                                            "date": "2025-05-23",
                                            "reason": "Beatae omnis optio dolor minus occaecati dolor.",
                                            "approvedAt": "2025-05-31 00:11:09"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/calendar/{date}": {
            "get": {
                "tags": [
                    "Calendar"
                ],
                "summary": "Get Calendar for Spesific Date",
                "description": "Get calendar data for a date",
                "operationId": "calendar.date",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "Date in YYYY-MM-DD format",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "2025-05-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": [
                                        {
                                            "name": "Test User",
                                            "reason": "Odio quibusdam suscipit nemo repellendus fugit.",
                                            "approvedAt": null
                                        },
                                        {
                                            "name": "Test User",
                                            "reason": "Beatae omnis optio dolor minus occaecati dolor.",
                                            "approvedAt": "2025-05-31 00:11:09"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "token_name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "default": "test@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "default": "password"
                                    },
                                    "token_name": {
                                        "type": "string",
                                        "default": "swagger"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "token": "1|5AujuSJSp03eIQNee0ZkyqxiSBBfGIO5RgpAG8Bc227437f5"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Authentication related endpoints"
        },
        {
            "name": "Calendar",
            "description": "Calendar related endpoints"
        }
    ]
}